+++ docker/sds-api/spine-directory-service/sds/request/routing_reliability_handler.py
@@ -15,24 +15,9 @@ from request.http_headers import HttpHeaders
 from request.tracking_ids_headers_reader import read_tracking_id_headers
 from utilities import timing, integration_adaptors_logger as log, mdc
 from utilities import config
+from utilities.constants import RELIABLE_SERVICES, FORWARD_RELIABLE_INTERACTIONS, FORWARD_EXPRESS_INTERACTIONS, FORWARD_RELIABLE_CORE_SPINE_SERVICE_INTERACTION, FORWARD_EXPRESS_CORE_SPINE_SERVICE_INTERACTION
 
 logger = log.IntegrationAdaptorsLogger(__name__)
-
-RELIABLE_SERVICES = [
-    'cc', 'ebs', 'ebsepr', 'ebsnpr', 'gp2gp', 'pat', 'itk', 'dis',
-    'ed', 'op', 'caf', 'adm', 'ooh', 'am', 'mh', 'nd', 'dir']
-FORWARD_RELIABLE_INTERACTIONS = [
-    'COPC_IN000001UK01', 'PRSC_IN040000UK08', 'PRSC_IN080000UK07', 'PRPA_IN010000UK07', 'PRPA_IN020000UK06',
-    'PRSC_IN050000UK06', 'PRSC_IN090000UK09', 'PRPA_IN030000UK08', 'PRSC_IN100000UK06', 'PRSC_IN070000UK08',
-    'PRSC_IN140000UK06', 'RCMR_IN010000UK05', 'RCMR_IN030000UK06', 'PRSC_IN130000UK07', 'PRSC_IN110000UK08',
-    'PRSC_IN060000UK06', 'PRSC_IN150000UK06', 'POLB_IN020006UK01', 'POLB_IN020005UK01', 'COMT_IN000004GB01',
-    'MCCI_IN010000UK13'
-]
-FORWARD_EXPRESS_INTERACTIONS = [
-    'PRSC_IN080000UK03', 'PRSC_IN040000UK03'
-]
-FORWARD_RELIABLE_CORE_SPINE_SERVICE_INTERACTION = 'urn:nhs:names:services:tms:ReliableIntermediary'
-FORWARD_EXPRESS_CORE_SPINE_SERVICE_INTERACTION = 'urn:nhs:names:services:tms:ExpressIntermediary'
 
 
 class RoutingReliabilityRequestHandler(BaseHandler, ErrorHandler):
@@ -65,13 +50,16 @@ class RoutingReliabilityRequestHandler(BaseHandler, ErrorHandler):
         
         if cpm_filter and cpm_filter[0] == CPM_FILTER_IDENTIFIER:
             ldap_results = await get_endpoint_from_cpm(org_code, service_id, party_key)
+            
+            logger.info("Obtained routing and reliability information. {ldap_results}",
+                        fparams={"ldap_results": ldap_results})
         else:
             ldap_results = await self.sds_client.get_mhs_details(org_code, service_id, party_key)
-        
-        logger.info("Obtained routing and reliability information. {ldap_results}",
-                    fparams={"ldap_results": ldap_results})
+            
+            logger.info("Obtained routing and reliability information. {ldap_results}",
+                        fparams={"ldap_results": ldap_results})
 
-        await self._handle_forward_reliable_results(ldap_results)
+            await self._handle_forward_reliable_results(ldap_results)
 
         base_url = f"{self.request.protocol}://{self.request.host}{self.request.path}/"
         full_url = unquote(self.request.full_url())
@@ -144,7 +132,7 @@ class RoutingReliabilityRequestHandler(BaseHandler, ErrorHandler):
     def _validate_query_params(self):
         query_params = self.request.arguments
         for query_param in query_params.keys():
-            if query_param not in [ORG_CODE_QUERY_PARAMETER_NAME, IDENTIFIER_QUERY_PARAMETER_NAME]:
+            if query_param not in [ORG_CODE_QUERY_PARAMETER_NAME, IDENTIFIER_QUERY_PARAMETER_NAME, CPM_FILTER]:
                 raise tornado.web.HTTPError(
                     status_code=400,
                     log_message=f"Illegal query parameter '{query_param}'")
@@ -156,6 +144,9 @@ class RoutingReliabilityRequestHandler(BaseHandler, ErrorHandler):
                 if query_param == IDENTIFIER_QUERY_PARAMETER_NAME \
                         and not query_param_value.startswith(f"{SERVICE_ID_FHIR_IDENTIFIER}|") \
                         and not query_param_value.startswith(f"{PARTY_KEY_FHIR_IDENTIFIER}|"):
+                    self._raise_invalid_identifier_query_param_error()
+                if query_param == CPM_FILTER and query_param_value.lower() != CPM_FILTER_IDENTIFIER:
+                    # Raise if not correct.
                     self._raise_invalid_identifier_query_param_error()
 
     @staticmethod